+++ stable_baselines/common/segment_tree.py
@@ -1,4 +1,18 @@
-import operator
+import numpy as np
+
+
+def unique(sorted_array):
+    """
+    More efficient implementation of np.unique for sorted arrays
+    :param sorted_array: (np.ndarray)
+    :return:(np.ndarray) sorted_array without duplicate elements
+    """
+    if len(sorted_array) == 1:
+        return sorted_array
+    left = sorted_array[:-1]
+    right = sorted_array[1:]
+    uniques = np.append(right != left, True)
+    return sorted_array[uniques]
 
 
 class SegmentTree(object):
@@ -8,7 +22,7 @@ class SegmentTree(object):
 
         https://en.wikipedia.org/wiki/Segment_tree
 
-        Can be used as regular array, but with two
+        Can be used as regular array that supports Index arrays, but with two
         important differences:
 
             a) setting item's value is slightly slower.
@@ -26,6 +40,7 @@ class SegmentTree(object):
         self._capacity = capacity
         self._value = [neutral_element for _ in range(2 * capacity)]
         self._operation = operation
+        self.neutral_element = neutral_element
 
     def _reduce_helper(self, start, end, node, node_start, node_end):
         if start == node_start and end == node_end:
@@ -61,19 +76,25 @@ class SegmentTree(object):
         return self._reduce_helper(start, end, 1, 0, self._capacity - 1)
 
     def __setitem__(self, idx, val):
-        # index of the leaf
-        idx += self._capacity
-        self._value[idx] = val
-        idx //= 2
-        while idx >= 1:
-            self._value[idx] = self._operation(
-                self._value[2 * idx],
-                self._value[2 * idx + 1]
+        # indexes of the leaf
+        idxs = idx + self._capacity
+        self._value[idxs] = val
+        if isinstance(idxs, int):
+            idxs = np.array([idxs])
+        # go up one level in the tree and remove duplicate indexes
+        idxs = unique(idxs // 2)
+        while len(idxs) > 1 or idxs[0] > 0:
+            # as long as there are non-zero indexes, update the corresponding values
+            self._value[idxs] = self._operation(
+                self._value[2 * idxs],
+                self._value[2 * idxs + 1]
             )
-            idx //= 2
+            # go up one level in the tree and remove duplicate indexes
+            idxs = unique(idxs // 2)
 
     def __getitem__(self, idx):
-        assert 0 <= idx < self._capacity
+        assert np.max(idx) < self._capacity
+        assert 0 <= np.min(idx)
         return self._value[self._capacity + idx]
 
 
@@ -81,9 +102,10 @@ class SumSegmentTree(SegmentTree):
     def __init__(self, capacity):
         super(SumSegmentTree, self).__init__(
             capacity=capacity,
-            operation=operator.add,
+            operation=np.add,
             neutral_element=0.0
         )
+        self._value = np.array(self._value)
 
     def sum(self, start=0, end=None):
         """
@@ -98,23 +120,34 @@ class SumSegmentTree(SegmentTree):
     def find_prefixsum_idx(self, prefixsum):
         """
         Find the highest index `i` in the array such that
-            sum(arr[0] + arr[1] + ... + arr[i - i]) <= prefixsum
+            sum(arr[0] + arr[1] + ... + arr[i - i]) <= prefixsum for each entry in prefixsum
 
         if array values are probabilities, this function
         allows to sample indexes according to the discrete
         probability efficiently.
 
-        :param prefixsum: (float) upperbound on the sum of array prefix
-        :return: (int) highest index satisfying the prefixsum constraint
+        :param prefixsum: (np.ndarray) float upper bounds on the sum of array prefix
+        :return: (np.ndarray) highest indexes satisfying the prefixsum constraint
         """
-        assert 0 <= prefixsum <= self.sum() + 1e-5
-        idx = 1
-        while idx < self._capacity:  # while non-leaf
-            if self._value[2 * idx] > prefixsum:
-                idx = 2 * idx
-            else:
-                prefixsum -= self._value[2 * idx]
-                idx = 2 * idx + 1
+        if isinstance(prefixsum, float):
+            prefixsum = np.array([prefixsum])
+        assert 0 <= np.min(prefixsum)
+        assert np.max(prefixsum) <= self.sum() + 1e-5
+        assert isinstance(prefixsum[0], float)
+
+        idx = np.ones(len(prefixsum), dtype=int)
+        cont = np.ones(len(prefixsum), dtype=bool)
+
+        while np.any(cont):  # while not all nodes are leafs
+            idx[cont] = 2 * idx[cont]
+            prefixsum_new = np.where(self._value[idx] <= prefixsum, prefixsum - self._value[idx], prefixsum)
+            # prepare update of prefixsum for all right children
+            idx = np.where(np.logical_or(self._value[idx] > prefixsum, np.logical_not(cont)), idx, idx + 1)
+            # Select child node for non-leaf nodes
+            prefixsum = prefixsum_new
+            # update prefixsum
+            cont = idx < self._capacity
+            # collect leafs
         return idx - self._capacity
 
 
@@ -122,9 +155,10 @@ class MinSegmentTree(SegmentTree):
     def __init__(self, capacity):
         super(MinSegmentTree, self).__init__(
             capacity=capacity,
-            operation=min,
+            operation=np.minimum,
             neutral_element=float('inf')
         )
+        self._value = np.array(self._value)
 
     def min(self, start=0, end=None):
         """